add paging for recyclerview

+++ app/src/main/java/me/pakhang/wanandroid/ui/home/Article.kt
@@ -47,7 +47,7 @@ package me.pakhang.wanandroid.ui.home
 }
 */
 data class Article(
-        val id: Long, val title: String, val author: String,
+        val id: Int, val title: String, val author: String,
         val superChapterName: String, val chapterName: String,
         val link: String
 )
